Time: 0 ms (100%), Space: 11.8 MB (78.65%) - LeetHub

+++ 1513-number-of-substrings-with-only-1s/1513-number-of-substrings-with-only-1s.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     const int mod=1e9+7;
 
-int solve(int n){
+   long long solve(long long n){
     return n*(n+1)/2;
 }
